requirements.txt
Django==4.0.3
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
django-allauth==0.5
dj-rest-auth==2.2.4
django-cors-headers==3.11
djangorestframework-simplejwt==5.1.0